agents/media_generator/requirements.txt
aiohttp
requests
Pillow
playwright
python-dotenv
